# Conflicts:
#	src/main/java/com/phpn/services/customer/CustomerService.java

+++ src/main/java/com/phpn/controllers/api/CustomerAPI.java
@@ -59,7 +59,7 @@ public class CustomerAPI {
 }
 
     @PostMapping("/delete/{id}")
-    public void deleteCustomerById(@PathVariable Long id) {
+    public void deleteCustomerById(@PathVariable Integer id) {
         customerService.deleteCustomer(id);
     }
